MDEV-31273: Precompute binlog checksums
Intermediate commit. Preparation for eliminating Log_event::checksum_alg.

Which checksum algorithm is used (if any) when writing an event does not
belong in the event, it is a property of the log being written to. Instead
introduce a client-only Log_event::read_checksum_alg to be able to print the
checksum read, and a Format_description_log_event::source_checksum_alg which
is the checksum algorithm (if any) to use when reading events from a log.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>