MDEV-31835 [WIP] Reduce calls to extra(HA_EXTRA_IGNORE_INSERT)
HA_EXTRA_IGNORE_INSERT, introduced in MDEV-515, is used to inform the
storage engine of INSERT IGNORE through the extra() method. These
calls are generally only needed at most once per statement, so we move
them out of write_record(), into the callsite, or grand callsite,
before loops invoking write_record(). The only places we have not done
so are select_insert::send_data(), where we move the call to just
before write_record(), and Delayed_insert::handle_inserts(), where the
dup and ignore are per row.

This will improve performance in some scenarios. For example, in `load
data infile ignore' to a partitioned table it will reduce the number
of calls by a factor of num_rows x num_partitions, and for a
non-partitioned table by a factor of num_rows.